ado2gh Factories
