+++ src/OctoshiftCLI.Tests/ado2gh/Services/TeamProjectsCsvGeneratorServiceTests.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using FluentAssertions;
 using Moq;
 using OctoshiftCLI.AdoToGithub;
+using OctoshiftCLI.AdoToGithub.Factories;
 using OctoshiftCLI.Services;
 using Xunit;
