+++ src/ado2gh/Commands/InventoryReportCommand.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.CommandLine;
 using Microsoft.Extensions.DependencyInjection;
+using OctoshiftCLI.AdoToGithub.Factories;
 using OctoshiftCLI.AdoToGithub.Handlers;
 using OctoshiftCLI.Commands;
 using OctoshiftCLI.Services;
